[OPIK-1988] metric date range select refactoring
