+++ apps/opik-frontend/src/components/pages-shared/traces/MetricDateRangeSelect/constants.ts
@@ -1,5 +1,5 @@
+import { DateRangeValue } from "@/components/shared/DateRangeSelect";
 import dayjs from "dayjs";
-import { DateRangeValue } from "@/components/shared/DateRangeSelect/DateRangeSelect";
 
 export const DEFAULT_METRICS_DATE_RANGE: DateRangeValue = {
   from: dayjs().subtract(29, "days").startOf("day").toDate(),
