+++ apps/opik-frontend/src/components/shared/DateRangeSelect/index.ts
@@ -1,0 +1,3 @@
+export { default as DateRangeSelect } from "./DateRangeSelect";
+export type { DateRangeValue, DateRangePreset } from "./types";
+export * from "./constants";
