+++ apps/opik-frontend/src/components/shared/DateRangeSelect/types.ts
@@ -1,0 +1,6 @@
+export type DateRangePreset = "today" | "last3days" | "lastWeek" | "lastMonth";
+
+export type DateRangeValue = {
+  from: Date;
+  to: Date;
+};
